End of the World
Given the world a finite length. The level is won once the chicken
reaches a certain point, any button pressed after that will start the
next level.
An interface as been created for the PlayScreen so that the world can
send messages back up the hierarchy (such as nextLevel()).
Also given the platforms varying length.